Clients now have a separate thread for listening.
Implemented a thread in the client side so they can Listen and Send messages at the same time.
Replaced userHandler by:
Listen method that creates a thread and listen for incoming messages from the sever.
Send method that awaits for the client to input a command and send it to the server.