Added the talk and read functionality
The server now detects the commands talk and read, as well as t and r, to be used for client-client communication.
Also cleaned up a lot of the code.
Any new commands that the server should detect just need to be added to the Switch in the run method of the threads.